graphkir/requirements.txt
numpy
pandas
pandas-stubs
scipy
scikit-learn
pysam
biopython
git+https://github.com/linnil1/pyHLAMSA
matplotlib
plotly
